Downgrade sharp version spec for Tools from ^0.25.2 to ^0.23.4
This work arounds the build failure on FreeBSD, which is caused by
the fact that sharp 0.25.x requires libvips 8.9.1 while the latest
FreeBSD package for vips is still 8.8.3.

Sharp 0.23.4 is the latest version which can be built with vips 8.8.x
and it is also the one being specified in CliClient/package.json.